+++ lib/types.ts
@@ -32,6 +32,11 @@ interface CommonOptions {
   color?: boolean;
   logLevel?: LogLevel;
   errorLimit?: number;
+
+  removeConsole?: boolean
+  removeDebugger?: boolean
+  debugTool?: string
+  removeDebugTool?: boolean
 }
 
 export interface BuildOptions extends CommonOptions {