+++ scripts/ts-type-tests.js
@@ -171,6 +171,10 @@ const tests = {
       },
       sourcefile: '',
       loader: 'ts',
+      removeConsole: false,
+      removeDebugger: false,
+      debugTool: '',
+      removeDebugTool: false,
     }).then(result => {
       let code: string = result.code;
       let map: string = result.map;
@@ -231,6 +235,10 @@ const tests = {
         sourcefile: '',
         loader: 'ts',
       },
+      removeConsole: false,
+      removeDebugger: false,
+      debugTool: '',
+      removeDebugTool: false,
       plugins: [
         {
           name: 'x',